+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Unreleased
 - Made various `enum` types non-exhaustive
 - Added `sys::VertexBufferUsage` enum and adjusted
   `sys::Context::set_vertex_buffer_data` to require an instance
+- Added `sys::Context::set_vertex_buffer_sub_data` method
 - Renamed `PrimitiveTy` trait to `BuiltinType`
 - Renamed `DrawType` enum to `Primitive`